dev (#1)
accept a paremeter for build mode (and parameterized cron job to run both options), one to build the latest version from the master branch and one that check out the most recent tag/release and builds that. can also accept the name of a tag to build. reads the version number from configure.ac.
Co-Authored-By: ben <ben@sudo.is>
Co-Committed-By: ben <ben@sudo.is>